+++ src/main/java/net/swordie/ms/scripts/ScriptManager.java
@@ -1662,7 +1662,7 @@ public interface ScriptManager extends Observer {
 
 	void spawnLotus (byte phase, byte difficulty);
 
-	void addStorageSlots(byte amount);
+	void addStorageSlots(int amount);
 
 	void addInventorySlotsByInvType(byte amount, byte type);
